Switched Handlers to use id() for IDs.
Handler.hid was removed in favor of using the value returned by id().
HandlerManager._counter is no longer necessary, so it was removed. HandlerManager.add_handler() now returns id(handler).

Differing objects with non-overlapping lifetimes may have equal id() values (see http://docs.python.org/3/library/functions.html#id ). It was decided that this would not be an issue as any objects that might store a Handler's ID would most likely be destroyed by the time the Handler was. This should probably be noted in the documentation.