First plugins; removed Handler.pattern
There was a TODO in connection.py saying to make a bunch of handlers into plugins.
So I did.

Also removed `Handler.pattern`. Dunno when I was thinking when I added that.
Regexes are /not/ how you parse these things. (Right?)

Also I fixed up a couple of redundant things.