Revamped HandlerManager; added sortedlist.py
HandlerManager now implements the mutable-set interface.
It also uses a defaultdict for the the event->sequence-of-handlers mapping,
and `SortedList`s (from sortedlist.py) for said sequence.

__repr__(), __str__(), __iter__(), add(), and remove() were rewritten, and
__contains__() and discard() methods were added. All the other mutable-set
methods were inherited from the ABC.

__repr__() now includes the class name.
__str__() now returns a proper human-readable summary of the manager.
__iter__() is now a generator.

The length is now kept as an instance variable instead of being computed, and
the handler_lists attribute was renamed to _handlers.